requirements.txt
loguru==0.7.3
pybrowsers==0.6.0
python-dotenv==1.0.1
selenium==4.28.1
webdriver-manager==4.0.2